+++ packages/main/config/postcss.bundles/postcss.config.js
@@ -1,7 +1,7 @@
 const postcssImport = require('postcss-import');
 const combineSelectors = require('postcss-combine-duplicated-selectors');
 const postcssCSStoESM = require('../../lib/postcss-css-to-esm/index.js');
-const postcssDerivedColors = require('../../lib/postcss-override-base-params/process-derived-colors');
+const postcssDerivedColors = require('../../lib/postcss-process-derived-colors/index');
 const cssnano = require('cssnano');
 const postcssAddFallback = require('../../lib/postcss-add-fallback/index.js');
